Configuración base tienda 16

+++ back/controllers/ConfigController.js
@@ -1,5 +1,6 @@
 'use strict';
 var config = require('../models/config');
+var fs = require('fs');
 
 const obtener_config_admin = async (req, res) => {
     if (req.user) {
@@ -18,39 +19,50 @@ const actualizar_config_admin = async (req, res) => {
     if (req.user) {
         if (req.user.role == "admin") {
             let data = req.body;
+            let reg;
 
-            if (req.files) {
-                // Si hay imagen
+            if (req.files && req.files.logo) {
+                console.log('si hay imagen');
                 var img_path = req.files.logo.path;
                 var name = img_path.split("\\");
-                var logo_name = name[2];
+                var logo_name = name[name.length - 1]; 
 
-                let reg = await config.findByIdAndUpdate({ _id: "68daa75d1e1062bf51932fa2" }, {
-                    categorias: data.categorias,
-                    titulo: data.titulo,
-                    logo: logo_name,
-                    serie: data.serie,
-                    correlativo: data.correlativo,
-                });
-
-                fs.stat("./uploads/configuraciones/" + reg.logo, function (err) {
-                    if (!err) {
-                        fs.unlink("./uploads/configuraciones/" + reg.logo, (err) => {
-                            if (err) throw err;
-                        });
+                reg = await config.findByIdAndUpdate(
+                    { _id: "68daa75d1e1062bf51932fa2" },
+                    {
+                        categorias: data.categorias,
+                        titulo: data.titulo,
+                        logo: logo_name,
+                        serie: data.serie,
+                        correlativo: data.correlativo,
                     }
-                });
+                );
+
+                if (reg.logo) {
+                    fs.stat("./uploads/configuraciones/" + reg.logo, function (err) {
+                        if (!err) {
+                            fs.unlink("./uploads/configuraciones/" + reg.logo, (err) => {
+                                if (err) throw err;
+                            });
+                        }
+                    });
+                }
 
             } else {
-                let reg = await config.findByIdAndUpdate({ _id: "68daa75d1e1062bf51932fa2" }, {
-                    categorias: data.categorias,
-                    titulo: data.titulo,
-                    serie: data.serie,
-                    correlativo: data.correlativo,
-                });
+                console.log('no hay imagen');
+                reg = await config.findByIdAndUpdate(
+                    { _id: "68daa75d1e1062bf51932fa2" },
+                    {
+                        categorias: data.categorias,
+                        titulo: data.titulo,
+                        serie: data.serie,
+                        correlativo: data.correlativo,
+                    }
+                );
             }
 
             res.status(200).send({ data: reg });
+
         } else {
             res.status(500).send({ message: "No autorizado" });
         }
@@ -58,7 +70,6 @@ const actualizar_config_admin = async (req, res) => {
         res.status(500).send({ message: "No autorizado" });
     }
 }
-
 
 module.exports = {
     actualizar_config_admin,
